requirements.txt
py-cord==2.0.0rc1
PyYAML==6.0